Updated notification service

+++ src/main/java/com/algolens/algo_lens/dtos/contest/UpcomingContestDTO.java
@@ -10,6 +10,6 @@ public record UpcomingContestDTO(
         String name,
         String type,
         int durationSeconds,
-        int startTimeSeconds,
+        long startTimeSeconds,
         int relativeTimeSeconds){
 }
